remove fallback remotes
From now on, I am going to assume that PBS Admin will be used with bytenet (and later, my own networking library based on bytenet once i am done with that). This simplifies implementation at the cost of making setup a little more difficult, but for most users, this will be done for them as pbs admin comes with rewrite.